// src/types/subscription.ts
export type PlanD = {
  planDuration: string;
  actualPrice: string;
  discountedPrice: string;
  totalAmount: string;
  discount: string;
  id: number;
  inputId: string;
};

export type TablePlansD = {
  label: string;
  icon: any;
  isIcon: boolean;
};

export type CoundownD = {
  label: string;
  value: number;
};